use init.lua from kickstart.nvim
This can be found here: https://github.com/nvim-lua/kickstart.nvim
The plan is to use this as a starting point for nvim-lspconfig
and adapt it to my need, which to be fair, would probably not require
a great deal of modification. Plus, as a vimscript and lua novice, configuring
in lua feels a lot more intuitive.